Refactor noise-refresh CRT formatter helpers
