+++ src/noise_refresh/mod.rs
@@ -1,2 +1,2 @@
-pub mod circuits;
-pub mod fhe_prf;
+pub mod step_fn_format;
+pub mod step_fn_prg;
